backend/requirements.txt
Django==2.1.7
pytz==2018.9
google-cloud-storage==1.16.0